+++ setup/owncloud.sh
@@ -10,7 +10,7 @@ source /etc/mailinabox.conf # load global vars
 apt_install \
 	dbconfig-common \
 	php5-cli php5-sqlite php5-gd php5-imap php5-curl php-pear php-apc curl libapr1 libtool libcurl4-openssl-dev php-xml-parser \
-	php5 php5-dev php5-gd php5-fpm memcached php5-memcache unzip
+	php5 php5-dev php5-gd php5-fpm memcached php5-memcached unzip
 
 apt-get purge -qq -y owncloud*
 
@@ -102,6 +102,7 @@ if [ ! -f $STORAGE_ROOT/owncloud/owncloud.db ]; then
       'arguments'=>array('{localhost:993/imap/ssl/novalidate-cert}')
     )
   ),
+  'memcache.local' => '\\OC\\Memcache\\Memcached',
   "memcached_servers" => array (
     array('localhost', 11211),
   ),
@@ -147,6 +148,21 @@ EOF
 	# settings and deletes the autoconfig.php file.
 	(cd /usr/local/lib/owncloud; sudo -u www-data php /usr/local/lib/owncloud/index.php;)
 fi
+
+# Update existing configuration files with changed settings that weren't included in
+# previous versions of Mail-in-a-Box. Use PHP to read the settings file, modify it,
+# and write out the new settings array.
+CONFIG_TEMP=$(/bin/mktemp)
+php <<EOF > $CONFIG_TEMP && mv $CONFIG_TEMP $STORAGE_ROOT/owncloud/config.php;
+<?php
+include("$STORAGE_ROOT/owncloud/config.php");
+\$CONFIG['memcache.local'] = '\\OC\\Memcache\\Memcached';
+echo "<?php\n\\\$CONFIG = ";
+var_export(\$CONFIG);
+echo ";";
+?>
+EOF
+chown www-data.www-data $STORAGE_ROOT/owncloud/config.php
 
 # Enable/disable apps. Note that this must be done after the ownCloud setup.
 # The firstrunwizard gave Josh all sorts of problems, so disabling that.